norm: removed unnescessary var from game_loop
